requirements.txt
streamlit
google-cloud-aiplatform
google-auth
requests
beautifulsoup4